requirements.txt
streamlit
openai
pydantic_ai
python-dotenv
requests
mcp[sse]